Tema_9_CLIENTE_API/requirements.txt
Django~=4.2.7
psycopg2-binary~=2.9.9
django-debug-toolbar~=4.2.0
django-bootstrap-v5~=1.0.11
django-bootstrap-icons~=0.8.6
django-jquery~=3.1.0
django-bootstrap-datepicker-plus~=5.0.4
django-staticfiles-moment~=2.0.0
requests~=2.31.0
django-environ~=0.11.2